requirements.txt
aiohttp
async-timeout
attrs
chardet
discord.py
idna
multidict
yarl
